Use percent unit if all metrics has the same unit
